initial code done, not compiling
