+++ http/src/main/scala/org/broadinstitute/dsde/workbench/leonardo/http/api/RuntimeV2Routes.scala
@@ -12,15 +12,16 @@ import de.heikoseeberger.akkahttpcirce.ErrorAccumulatingCirceSupport._
 import io.circe.Decoder
 import io.opencensus.scala.akka.http.TracingDirective.traceRequestForService
 import org.broadinstitute.dsde.workbench.leonardo.config.RefererConfig
-import org.broadinstitute.dsde.workbench.leonardo.http.service.AzureService
+import org.broadinstitute.dsde.workbench.leonardo.http.service.RuntimeV2Service
 import org.broadinstitute.dsde.workbench.model.UserInfo
 import org.broadinstitute.dsde.workbench.openTelemetry.OpenTelemetryMetrics
 import JsonCodec._
+import RuntimeRoutesCodec._
 import com.azure.core.management.Region
 import com.azure.resourcemanager.compute.models.VirtualMachineSizeTypes
 
 class RuntimeV2Routes(saturnIframeExtentionHostConfig: RefererConfig,
-                      azureService: AzureService[IO],
+                      runtimeV2Service: RuntimeV2Service[IO],
                       userInfoDirectives: UserInfoDirectives)(
   implicit metrics: OpenTelemetryMetrics[IO]
 ) {
@@ -40,6 +41,7 @@ class RuntimeV2Routes(saturnIframeExtentionHostConfig: RefererConfig,
                     listRuntimesHandler(
                       userInfo,
                       None,
+                      None,
                       params
                     )
                   )
@@ -54,7 +56,8 @@ class RuntimeV2Routes(saturnIframeExtentionHostConfig: RefererConfig,
                       complete(
                         listRuntimesHandler(
                           userInfo,
-                          Some(cloudContext),
+                          Some(workspaceId),
+                          Some(CloudProvider.Azure),
                           params
                         )
                       )
@@ -128,10 +131,10 @@ class RuntimeV2Routes(saturnIframeExtentionHostConfig: RefererConfig,
       ctx <- ev.ask[AppContext]
 
       jobUUID = WsmJobId(s"create-${runtimeName.asString}")
-      apiCall = azureService.createRuntime(userInfo, runtimeName, workspaceId, req, jobUUID)
-      _ <- metrics.incrementCounter("createAzureRuntime")
+      apiCall = runtimeV2Service.createRuntime(userInfo, runtimeName, workspaceId, req, jobUUID)
+      _ <- metrics.incrementCounter("createRuntimeV2")
       _ <- ctx.span.fold(apiCall)(span =>
-        spanResource[IO](span, "createAzureRuntime")
+        spanResource[IO](span, "createRuntimeV2")
           .use(_ => apiCall)
       )
     } yield StatusCodes.Accepted: ToResponseMarshallable
@@ -141,10 +144,10 @@ class RuntimeV2Routes(saturnIframeExtentionHostConfig: RefererConfig,
   ): IO[ToResponseMarshallable] =
     for {
       ctx <- ev.ask[AppContext]
-      apiCall = azureService.getRuntime(userInfo, runtimeName, workspaceId)
-      _ <- metrics.incrementCounter("getAzureRuntime")
+      apiCall = runtimeV2Service.getRuntime(userInfo, runtimeName, workspaceId)
+      _ <- metrics.incrementCounter("getRuntimeV2")
       resp <- ctx.span.fold(apiCall)(span =>
-        spanResource[IO](span, "getAzureRuntime")
+        spanResource[IO](span, "getRuntimeV2")
           .use(_ => apiCall)
       )
     } yield StatusCodes.OK -> resp: ToResponseMarshallable
@@ -157,10 +160,10 @@ class RuntimeV2Routes(saturnIframeExtentionHostConfig: RefererConfig,
   ): IO[ToResponseMarshallable] =
     for {
       ctx <- ev.ask[AppContext]
-      apiCall = azureService.updateRuntime(userInfo, runtimeName, workspaceId, req)
-      _ <- metrics.incrementCounter("updateAzureRuntime")
+      apiCall = runtimeV2Service.updateRuntime(userInfo, runtimeName, workspaceId, req)
+      _ <- metrics.incrementCounter("updateRuntimeV2")
       _ <- ctx.span.fold(apiCall)(span =>
-        spanResource[IO](span, "updateAzureRuntime")
+        spanResource[IO](span, "updateRuntimeV2")
           .use(_ => apiCall)
       )
     } yield StatusCodes.Accepted: ToResponseMarshallable
@@ -170,13 +173,29 @@ class RuntimeV2Routes(saturnIframeExtentionHostConfig: RefererConfig,
   ): IO[ToResponseMarshallable] =
     for {
       ctx <- ev.ask[AppContext]
-      apiCall = azureService.deleteRuntime(userInfo, runtimeName, workspaceId)
-      _ <- metrics.incrementCounter("deleteAzureRuntime")
+      apiCall = runtimeV2Service.deleteRuntime(userInfo, runtimeName, workspaceId)
+      _ <- metrics.incrementCounter("deleteRuntimeV2")
       _ <- ctx.span.fold(apiCall)(span =>
-        spanResource[IO](span, "deleteRuntime")
+        spanResource[IO](span, "deleteRuntimeV2")
           .use(_ => apiCall)
       )
     } yield StatusCodes.Accepted: ToResponseMarshallable
+
+  private[api] def listRuntimesHandler(userInfo: UserInfo,
+                                       workspaceId: Option[WorkspaceId],
+                                       cloudProvider: Option[CloudProvider],
+                                       params: Map[String, String])(
+                                        implicit ev: Ask[IO, AppContext]
+                                      ): IO[ToResponseMarshallable] =
+    for {
+      ctx <- ev.ask[AppContext]
+      apiCall = runtimeV2Service.listRuntimes(userInfo, workspaceId, cloudProvider, params)
+      _ <- metrics.incrementCounter("listRuntimeV2")
+      resp <- ctx.span.fold(apiCall)(span =>
+        spanResource[IO](span, "listRuntimeV2")
+          .use(_ => apiCall)
+      )
+    } yield StatusCodes.OK -> resp: ToResponseMarshallable
 
   implicit val createAzureDiskReqDecoder: Decoder[CreateAzureDiskRequest] =
     Decoder.forProduct4("labels", "name", "size", "diskType")(CreateAzureDiskRequest.apply)
